refactor(DatePriceResponse): - maple 패키지로부터 config 패키지 분리 - static 상수 관리 클래스 변경
