+++ src/main/java/com/mmt/tracker/config/MapleApiClientConfiguration.java
@@ -1,0 +1,14 @@
+package com.mmt.tracker.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class MapleApiClientConfiguration {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
