--- src/main/java/com/mmt/tracker/maple/client/MapleApiConfig.java
@@ -1,29 +1,0 @@
-package com.mmt.tracker.maple.client;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class MapleApiConfig {
-
-    @Value("${maple.api.key}")
-    private String apiKey;
-
-    @Value("${maple.api.base-url}")
-    private String baseUrl;
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-
-    public String getApiKey() {
-        return apiKey;
-    }
-
-    public String getBaseUrl() {
-        return baseUrl;
-    }
-}
