+++ src/main/resources/application.properties
@@ -12,9 +12,10 @@ spring.config.import=classpath:application-secret.properties
 
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=update
 
 spring.datasource.url=jdbc:h2:file:./h2db/tracker
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
+
